+++ types/core.d.ts
@@ -1,14 +1,7 @@
-import { Atomico, DOMProps, JSXElements, JSXProps, Nullable } from "./dom";
+import { Atomico, JSXElements, JSXProps, Nullable } from "./dom";
 
-import {
-    FillObject,
-    SchemaProps,
-    ConstructorType,
-    TypeToConstructor,
-    SchemaInfer,
-} from "./schema";
+import { FillObject, TypeToConstructor } from "./schema";
 
-import { Sheets } from "./css";
 import { VNodeKeyTypes, VNode } from "./vnode";
 import * as Hooks from "./hooks";
 
@@ -16,6 +9,7 @@ export { DOMEvent, DOMListener, JSXElement } from "./dom";
 export { css, Sheet, Sheets } from "./css";
 export { html } from "./html";
 export * from "./hooks";
+export { c, Props, Component, Host } from "./component";
 
 /**
  * Identify whether a node in the list belongs to a fragment marker instance
@@ -34,159 +28,7 @@ export interface Ref<CurrentTarget = any> extends FillObject {
         : CurrentTarget;
 }
 
-export type Host<Meta> = VNode<any>;
-
-export type SyntheticProps<Props> = {
-    [Prop in keyof Props]: Prop extends `on${string}`
-        ? {
-              type: Function;
-              value: (event: Props[Prop]) => any;
-          }
-        : {
-              type: Function;
-              value: Props[Prop];
-          };
-};
-
-export type SyntheticMetaProps<Meta> = {
-    [Prop in keyof Meta]?: Prop extends `on${string}`
-        ? (event: Meta[Prop]) => any
-        : Meta[Prop];
-};
-
 export type Type<Types> = TypeToConstructor<Types> & { meta?: Types };
-
-/**
- * Infer the types from `component.props`.
- 
- * ```tsx
- * function component({value}: Props<typeof component.props >){
- *      return <host/>
- * }
- *
- * component.props = {value:Number}
- * ```
- */
-type GetProps<P> = P extends {
-    readonly "##props"?: infer P;
-}
-    ? P
-    : P extends { props: SchemaProps }
-    ? GetProps<P["props"]>
-    : {
-          [K in keyof P]?: P[K] extends {
-              type: infer T;
-              value: infer V;
-          }
-              ? FunctionConstructor extends T
-                  ? V
-                  : V extends () => infer T
-                  ? T
-                  : V
-              : P[K] extends { type: infer T }
-              ? ConstructorType<T>
-              : ConstructorType<P[K]>;
-      };
-
-type ReplaceProps<P, Types> = {
-    [I in keyof P]?: I extends keyof Types ? Types[I] : P;
-};
-
-/**
- * Infers the props from the component's props object, example:
- * ### Syntax
- * ```tsx
- * const myProps = { message: String }
- * Props<typeof MyProps>;
- * // {message: string}
- * ```
- * ### Usage
- * You can use the `Prop` type on components, objects or constructors, example:
- * ```tsx
- * function component({message}: Props<typeof component>){
- *  return <host></host>
- * }
- *
- * component.props = {message: String}
- * ```
- *
- * ### Advanced use
- *
- * It also allows to replace types of those already inferred, example:
- * ```tsx
- * Props<typeof MyProps, {message: "hello"|"bye bye"}>;
- * // {message?: "hello"|"bye bye"}
- *
- * ```
- */
-export type Props<P, Types = null> = Types extends null
-    ? GetProps<P>
-    : ReplaceProps<GetProps<P>, Types>;
-
-/**
- * metaProps allow to hide the props assigned by Component<props>
- */
-interface MetaProps<props> {
-    readonly "##props"?: props;
-}
-
-/**
- * The MetaComponent type allows to identify as
- * validate types generated by Component<props>
- */
-interface MetaComponent {
-    (props: any): any;
-    props: MetaProps<any>;
-    styles?: Sheets;
-}
-
-export type Component<props = null, meta = any> = props extends null
-    ? {
-          (props: FillObject): Host<meta>;
-          props?: SchemaProps;
-          styles?: Sheets;
-      }
-    : {
-          (props: DOMProps<props>): Host<meta>;
-          props: SchemaInfer<props> &
-              MetaProps<
-                  meta extends null ? props : props & SyntheticMetaProps<meta>
-              >;
-          styles?: Sheets;
-      };
-
-export type CreateElement<C, Base, CheckMeta = true> = CheckMeta extends true
-    ? C extends (props: any) => Host<infer Meta>
-        ? CreateElement<C & { props: SyntheticProps<Meta> }, Base, false>
-        : CreateElement<C, Base, false>
-    : C extends { props: infer P }
-    ? Atomico<Props<P>, Base>
-    : Atomico<{}, Base>;
-
-/**
- * Create the customElement to be declared in the document.
- * ### Usage
- * ```js
- * import {c} from "atomico";
- *
- * function myComponent(){
- *     return <host></host>
- * }
- *
- * customElements.define("my-component", c(myComponent));
- * ```
- * @todo Add a type setting that doesn't crash between JS and template-string.
- */
-export function c<
-    T extends typeof HTMLElement,
-    C extends Component | MetaComponent
->(component: C, BaseElement?: T): CreateElement<C, T>;
-
-export namespace h.JSX {
-    interface IntrinsicElements extends JSXElements {
-        [tagName: string]: any;
-    }
-}
 
 /**
  * function-pragma, create the vnode
